BackEnd/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
rq==1.15.1
redis==4.6.0
Werkzeug==2.3.7

